feat: decompose components
